requirements.txt
google-generativeai==0.4.1
pillow==10.2.0
SpeechRecognition==3.10.1
pyttsx3==2.90
PyAudio==0.2.14